+++ CHANGELOG.md
@@ -6,10 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 ### Updated
-- update links in QA to point internally for reverse proxy (DR-3237)
+
+- Updated links in QA to point internally for reverse proxy (DR-3237)
+- Updated how env vars are read for New Relic Browser implementation (DR-3235)
 
 ## [0.1.15] 2024-10-10
+
 ### Updated
 
 - Updated 500 and 404 error page designs, adding link to open feedback box (DR-3203)