+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Updated
 
-- update links in QA to point internally for reverse proxy (DR-3237)
-- Update cards to use Next Image (DR-3056)
+- Updated cards to use Next Image (DR-3056)
+- Updated links in QA to point internally for reverse proxy (DR-3237)
+- Updated how env vars are read for New Relic Browser implementation (DR-3235)
 
 ## [0.1.15] 2024-10-10